pyproject.toml
[tool.poetry]
name = "fillableRx"
version = "0.1.0"
description = ""
authors = ["Kevin Guerrero <guerrk83z@gmail.com>"]
readme = "README.md"
packages = [{include = "fillableRx"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.5"
django = "^4.1.3"
django-environ = "^0.9.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.0.1"}
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-twilio = "^0.14.3"
python-dotenv = "^0.21.0"
django-extensions = "^3.2.1"
django-on-heroku = "^1.1.2"
gunicorn = "^20.1.0"
python-decouple = "^3.7"
django-csp = "^3.7"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.5.1"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
